chore: include types into context file
