+++ src/context.tsx
@@ -1,7 +1,6 @@
 import { buildDecafClient, DecafClient } from '@decafhub/decaf-client';
 import Cookies from 'js-cookie';
 import React, { useContext } from 'react';
-import { DecafContextType, Principal, PublicConfig } from 'types';
 
 export const DecafContext = React.createContext<DecafContextType>({
   client: undefined as unknown as DecafClient,
@@ -49,4 +48,57 @@ export function getAuthenticatedDecafClient(): DecafClient | undefined {
 
   // Check token, build client and return:
   return token ? buildDecafClient('', { token }) : undefined;
+}
+
+export interface DecafContextType {
+  client: DecafClient;
+  me: Principal;
+  publicConfig: PublicConfig;
+}
+export interface Role {
+  code: string;
+  name: string;
+}
+
+export interface Team {
+  id: number;
+  name: string;
+}
+
+export interface Principal {
+  id: number;
+  guid: string;
+  username: string;
+  fullname: string;
+  first_name: string;
+  last_name: string;
+  email?: string;
+  mobile?: string;
+  active: boolean;
+  roles: Role[];
+  teams: Team[];
+  internal: boolean;
+  external: boolean;
+  privileged: boolean;
+}
+
+export interface PublicConfig {
+  /** company short name */
+  shortname: string;
+  /** company full name */
+  legalname: string;
+  /** company web site */
+  website: string;
+  /** logo url */
+  logo: string;
+  /** terms and conditions */
+  tnc: string;
+  /** zendeks code */
+  zendesk?: string;
+  /** google analytics code */
+  googleax?: string;
+  /** one-time password feature? */
+  otp: null;
+  /** password reset feature should be enabled or not */
+  pwdreset: true;
 }
