+++ src/index.tsx
@@ -1,6 +1,6 @@
-import ZendeskWidget from './ZendeskWidget';
 import { useDecaf } from './context';
 import DecafApp from './DecafApp';
 import DecafSpinner from './DecafSpinner';
+import ZendeskWidget from './ZendeskWidget';
 
 export { DecafApp, DecafSpinner, useDecaf, ZendeskWidget };
